Updating slampy

+++ slampy/sync.py
@@ -13,12 +13,49 @@ import threading
 import hashlib
 from threading import Timer
 
+"""
+how to generate new Google Drive API credentials:
+
+- Follow this https://pythonhosted.org/PyDrive/quickstart.html to generate credentials
+- you will download a JSON file 
+- Extract ClientId and Client Secret from the file
+- Create a client_settings.yaml with the content
+
+```
+client_config_backend: settings
+client_config:
+  client_id: "<REPLACE_HERE_AS_NEEDED>"
+  client_secret: "<REPLACE_HERE_AS_NEEDED>"
+
+save_credentials: True
+save_credentials_backend: file
+save_credentials_file: client_secrets.json
+
+get_refresh_token: True
+
+oauth_scope:
+  - https://www.googleapis.com/auth/drive.file
+  - https://www.googleapis.com/auth/drive.install
+```
+
+"""
+# 
+# basically and from that you extract the info to filla `client_settings.yaml`
 from pydrive.auth  import GoogleAuth
 from pydrive.drive import GoogleDrive
 
+ThisDir = os.path.dirname(os.path.realpath(__file__))
+
 T1=datetime.datetime.now() 
+
+# memory card -> local directory
 LOCAL_DIR="c:/visoar_files"
+
+# local diretory to google bucket name
 BUCKET_NAME="/visoar_files"
+
+# local directory to usb backup drive
+BACKUP_DIR="{}/visoar_files"
 
 # ///////////////////////////////////////////////
 class LogFile:
@@ -63,7 +100,7 @@ class Utils:
 			print(ret)
 			return ret
 				
-		fixed,external=[],[]
+		ret=[]
 		names=getInfo("name")
 		descriptions=getInfo("description");descriptions=descriptions + ([''] * (len(names)-len(descriptions)))
 		ids=getInfo("VolumeSerialNumber");ids=ids + (['0'] * (len(names)-len(ids)))
@@ -72,13 +109,22 @@ class Utils:
 		for name,description,id in zip(names,descriptions,ids):
 			description=description.lower()
 			print(name,description,id)
-			if "local fixed" in description or "cd-rom" in description:
-				fixed.append((name,id))
-			else:
-				external.append((name,id))
-		log.print("Fixed drives",fixed)
-		log.print("External drives",external)
-		return external	
+			if id=='0' or "local fixed" in description or "cd-rom" in description:
+				continue
+			ret.append((name,id))
+		log.print("External drives",ret)
+		return ret	
+	
+	# EjectUSBDrive (BROKEN!
+	@staticmethod
+	def EjectUSBDrive(drive):
+		tmp_file = open('tmp.ps1','w')
+		tmp_file.write("""$driveEject = New-Object -comObject Shell.Application\n""")
+		tmp_file.write("""$driveEject.Namespace(17).ParseName("{}").InvokeVerb("Eject")\n""".format(drive))
+		tmp_file.close()
+		process = subprocess.Popen(['powershell.exe', '-ExecutionPolicy','Unrestricted','.\\tmp.ps1'],stdout=sys.stdout)
+		process.communicate()
+		time.sleep(2)
 	
 	# ComputeMd5
 	@staticmethod
@@ -95,36 +141,44 @@ class Utils:
 class CopyLocalFiles:
 	
 	# constructor
-	def __init__(self):
-		pass
+	def __init__(self,dst_dir):
+		self.dst_dir=dst_dir
 	
 	# copyFile
-	def copyFile(self,src,dst,use_checksum=True):
+	def copyFile(self,src,dst):
+		
+		dst=os.path.join(self.dst_dir,dst).replace("\\","/")
+		
 		os.makedirs(os.path.dirname(dst), exist_ok=True) # create parent directory
-		if use_checksum:
-			l1=Utils.ComputeMd5(src)
-			l2=Utils.ComputeMd5(dst)
-			if l1==l2:
-				log.print("File {} already on destination. No need to copy again".format(src),"md5",l1)
-				return True
-		shutil.copy2(src, dst)	# copy2 preserve metadata
-		log.print("Copied {} file to local {} md5({})".format(src, dst,l1))	
-		return True		
 		
-	# removeFile
-	def removeFile(self,src):
-		os.remove(src)
+		l1=Utils.ComputeMd5(src)
+		l2=Utils.ComputeMd5(dst)
+		if l1==l2:
+			log.print("File {} already on destination. No need to copy again".format(src),"md5",l1)
+			return
 		
+		# copy2 preserve metadata
+		shutil.copy2(src, dst)	
+
+		# verify the new data
+		l2=Utils.ComputeMd5(dst)
+		if l1!=l2:
+			raise Exception("MD5 check failed")
+			
+		log.print("CopyLocalFiles: copied {} file to local {} md5({})".format(src, dst,l1))	
+		
+		
+
 # //////////////////////////////////////////////////////////
 class CopyFilesToGoogleDrive:
 	
-	settings_file='client_settings.yaml'
+	settings_file=os.path.join(ThisDir,'client_settings.yaml')
 	
 	# constructor
-	def __init__(self):
+	def __init__(self,bucket_name):
+		self.bucket_name=bucket_name
 		self.auth=GoogleAuth(settings_file=CopyFilesToGoogleDrive.settings_file)
 		self.drive=GoogleDrive(self.auth)	
-
 		self.root={'title':'root','id':'root'}
 		self.folders={} # caching the folders
 
@@ -168,8 +222,9 @@ class CopyFilesToGoogleDrive:
 		return remote_files[0]['md5Checksum'] if remote_files else None
 
 	# copyFile
-	def copyFile(self,src,dst):
-
+	def copyFile(self,src, dst):
+		
+		dst=(self.bucket_name + "/" + dst).replace("\\","/")
 		parent, child =dst.rsplit('/',maxsplit=1)
 		parent=self.createFolder(parent)
 
@@ -178,28 +233,43 @@ class CopyFilesToGoogleDrive:
 		
 		if l1 == l2:
 			log.print("File {} already on Google drive. No need to copy again md5({})".format(src,l1))
-			return True
+			return 
 
 		file = self.drive.CreateFile({'title': child, 'parents': [{'id': parent['id']}]})
 		file.SetContentFile(src)
-		file.Upload() # this should raise an error?
-		log.print("Uploaded {} file to Google drive md5({})".format(src,l1))
-		return True
+		file.Upload() 
+		
+		if file['title'] != child:
+			raise Exception("uplod of file failed")
+			
+		if file['md5Checksum'] !=l1:
+			raise Exception("MD5 checksum failed")
+			
+		log.print("CopyFilesToGoogleDrive: uploaded {} file to Google drive md5({}). MD5 check ok".format(src,l1))
 	
+# //////////////////////////////////////////////////////////
+class CompoundCopyFiles:
 	
-	# removeFile
-	def removeFile(self,src):
-		os.remove(src)	
+	# ___init__
+	def __init__(self):
+		self.instances=[]
+		
+	# copyFile
+	def copyFile(self,src,dst):
+		for it in self.instances:
+			it.copyFile(src,dst)
+		
+		# no need to keep it around
+		if os.path.abspath(src)!=os.path.abspath(log.filename):
+			os.remove(src)
 	
 # //////////////////////////////////////////////////////////
 class SyncFiles:
 	
 	# constructor
-	def __init__(self,src_dir,dst_dir,copier=None,purge=False, progress=None):
+	def __init__(self,src_dir,copier=None,progress=None):
 		self.src_dir=src_dir
-		self.dst_dir=dst_dir
 		self.copier=copier
-		self.purge=purge
 		self.progress=progress
 		
 	# list all files in a directory
@@ -214,13 +284,11 @@ class SyncFiles:
 		
 		log.print("Syncing folders")
 		log.print("source directory", self.src_dir)
-		log.print("target directory", self.dst_dir)
 		
 		t1 = time.time()
 		
 		nfiles,tot_bytes=0,0
 		
-		dst_files=[]
 		src_files=self.findFiles(self.src_dir)
 		
 		# make sure it's the last one (otherwise I will lose some logs)
@@ -234,9 +302,8 @@ class SyncFiles:
 		for I,src in enumerate(src_files):
 
 			# do not user normalize path
-			dst=self.dst_dir + "/" + os.path.relpath(src, self.src_dir)
-			dst=dst.replace("\\","/") 
-			
+			dst=os.path.relpath(src, self.src_dir).replace("\\","/") 
+
 			file_size=Utils.GetFileSize(src)
 			
 			log.print("%d/%d" % (I,N),"Doing",src,dst,"...")
@@ -244,34 +311,27 @@ class SyncFiles:
 
 			nfiles+=1
 			tot_bytes+=file_size 
-			dst_files.append(dst)
 			
 			kb=tot_bytes/(1024)
 			sec=time.time()-t1
 			kb_sec=int(kb/sec) if sec else 0
-			log.print("%d/%d" % (I,N),"Done",src,dst,"{}KB".format(int(file_size/1024)),"{}KB/sec".format(kb_sec),"KB({})".format(kb))
+			log.print("%d/%d" % (I,N),"Done",src, dst,"{}KB".format(int(file_size/1024)),"{}KB/sec".format(kb_sec),"KB({})".format(kb))
 					
 			# generator for a workflow
 			if self.progress:
 				self.progress(I,N, src, dst, kb_sec)
-			
-			# purge source file only if equals
-			if self.purge:
-				self.copier.removeFile(src)
-				log.print("%d/%d" % (I,N),"rm",src)
 				
 			time.sleep(0)
 					
-		# try to remove all empty subdirectories
-		if self.purge and not self.findFiles(src_dir):
-			for it in os.listdir(src_dir):
-				shutil.rmtree(os.path.join(src_dir, it), ignore_errors=True)
+		# try to remove all empty subdirectories 
+		for sub in [os.path.join(self.src_dir,it) for it in os.listdir(self.src_dir)]:
+			if os.path.isdir(sub) and not self.findFiles(sub):
+				shutil.rmtree(sub, ignore_errors=True)
 			
 		kb=int(tot_bytes/1024)
 		sec=time.time()-t1
 		kb_sec=int(kb/sec) if sec else 0
 		log.print("All done in","%.2fsec" % sec, "{}KB".format(kb),"{}KB/sec".format(kb_sec))
-		return nfiles,tot_bytes,dst_files
 
 
 # ///////////////////////////////////////////////////////////
@@ -314,45 +374,68 @@ class VisoarMoveDataFromCardWidget(QWidget):
 			log.print("Got arguments", sys.argv)
 		
 		self.setWindowTitle("Sync files")
-		self.resize(600,200)
+		self.resize(400,200)
 		self.sensors={}
 		self.createGui()
+		
+	# checkForUpdates
+	def checkForUpdates(self):
+		print("Checking for updates")	
+
+		import git
+		g = git.Git(os.path.join(ThisDir,".."))
+		xx=g.pull('origin','master')
+		print(xx)
+		#message="Software has been updated. You NEED TO RESTART"
+		#log.print(message)
+		#QMessageBox.about(self, "Error", message)
+		
+		
+	def separator(self):
+		line = QLabel(" ")
+		#line.setFrameShape(QFrame.HLine)
+		#line.setFrameShadow(QFrame.Sunken)
+		return line
 		
 	#  createGui
 	def createGui(self):
 		
 		main_layout = QVBoxLayout()
+		self.setLayout(main_layout)
 		
+		main_layout.addLayout(self.hlayout([
+			self.createButton('Refresh GUI',callback=self.refreshButtons),
+			self.createButton('Check for updates',callback=self.checkForUpdates),
+		]))
 		
-		self.field_name=QLineEdit()
-		self.field_name.setText("myfield")
+		main_layout.addWidget(self.separator())
 		
 		self.prog=QLineEdit();self.prog.setEnabled(False)
+		main_layout.addWidget(QLabel('Current'))
+		main_layout.addWidget(self.prog)		
+		
 		self.src=QLineEdit();self.src.setEnabled(False)
+		main_layout.addWidget(QLabel('Source'))
+		main_layout.addWidget(self.src)
+				
 		self.dst=QLineEdit();self.dst.setEnabled(False)
 		self.kb_sec=QLineEdit();self.kb_sec.setEnabled(False)
+		main_layout.addWidget(QLabel('Destination'))
+		main_layout.addWidget(self.dst)		
 		
 		self.progress = QProgressBar(self)
 		self.progress.setValue(0)	
-			
-		main_layout.addWidget(QLabel('Field name'))
-		main_layout.addWidget(self.field_name)
-		main_layout.addWidget(QLabel('Current'))
-		main_layout.addWidget(self.prog)
-		main_layout.addWidget(QLabel('Source'))
-		main_layout.addWidget(self.src)
-		main_layout.addWidget(QLabel('Destination'))
-		main_layout.addWidget(self.dst)
 		main_layout.addWidget(QLabel('KB/sec'))
-		main_layout.addWidget(self.kb_sec)
+		main_layout.addWidget(self.kb_sec)		
+			
 		main_layout.addWidget(QLabel('Progression'))
 		main_layout.addWidget(self.progress)
 		
-		self.buttons_layout=QHBoxLayout()
-		main_layout.addLayout(self.buttons_layout)
-		self.setLayout(main_layout)
+		main_layout.addWidget(self.separator())
 		
-		self.refreshButtons()
+		main_layout.addLayout(self.refreshButtons())
+
+		
 		
 	# hlayout
 	def hlayout(self,items):
@@ -388,27 +471,56 @@ class VisoarMoveDataFromCardWidget(QWidget):
 	# 	refreshButtons
 	def refreshButtons(self):
 		
+		if not "buttons_layout" in dir(self):
+			self.buttons_layout=QVBoxLayout()
+		
 		self.buttons=[]
 		self.clearLayout(self.buttons_layout)
 		
-		for drive, id in Utils.GetUsbDrives():
-			
+		# assign unique usb IDs
+		usb_drives=Utils.GetUsbDrives()
+		for drive, id in usb_drives:
 			if not id in self.sensors:
-				self.sensors[id]=len(self.sensors)
-				
-			button=self.createButton('Dump USB {}'.format(drive),callback=lambda : self.copyLocalFiles(drive, self.sensors[id]))
-			self.buttons_layout.addWidget(button)
-			self.buttons.append(button)
-				
-		if os.path.isfile(CopyFilesToGoogleDrive.settings_file):
-			button=self.createButton('Upload to gdrive',callback=self.copyFilesToGoogleDrive)
-			self.buttons_layout.addWidget(button)
-		self.buttons.append(button)
+				self.sensors[id]=len(self.sensors)		
+
 			
-		button=self.createButton('Refresh',callback=self.refreshButtons)
-		self.buttons_layout.addWidget(button)
-		self.buttons.append(button)
-		
+		if usb_drives:
+			
+			hlayout=QHBoxLayout()
+			self.buttons_layout.addLayout(hlayout)
+			
+			hlayout.addWidget(QLabel("Dump memory card"))
+			
+			field_name=QLineEdit()
+			field_name.setText("myfield")
+			hlayout.addWidget(QLabel('Field name'))
+			hlayout.addWidget(field_name)
+
+			for drive, id in usb_drives:
+				button=self.createButton('Dump USB {}'.format(drive),callback=lambda : self.dumpMemoryCard(drive, field_name.text(), self.sensors[id]))
+				hlayout.addWidget(button)
+				self.buttons.append(button)
+				
+		# google
+		if os.path.isfile(CopyFilesToGoogleDrive.settings_file):
+			
+			hlayout=QHBoxLayout()
+			self.buttons_layout.addLayout(hlayout)			
+			
+			hlayout.addWidget(QLabel("Google Drive"))
+			
+			button=self.createButton('Upload',callback=lambda: self.copyToCloud(drive=None))
+			hlayout.addWidget(button)
+			self.buttons.append(button)
+			
+			# also add the option to copy the data to external USB
+			for drive, id in usb_drives:
+				button=self.createButton('Upload and copy to {}'.format(drive),callback=lambda : self.copyToCloud(drive=drive))
+				hlayout.addWidget(button)
+				self.buttons.append(button)
+				
+		return self.buttons_layout
+				
 	# createButton
 	def createButton(self,text,callback=None):
 		ret=QPushButton(text)
@@ -441,51 +553,73 @@ class VisoarMoveDataFromCardWidget(QWidget):
 	def threadFinished(self):
 		
 		self.setRunning(False)
+		
 		if self.worker.error_message:
 			message="Error copying files. Please retry ({})".format(self.worker.error_message)
 			log.print(message)
 			QMessageBox.about(self, "Error", message)
 		else:
-			message="Finished Dumping memory card"
+			message="Finished syncing files"
 			log.print(message)
-
+	
 	# runSyncInBackground
-	def runSyncInBackground(self,sync):
+	def runSyncInBackground(self,sync,finished=None):
 		self.thread = QThread()
 		self.worker = MyWorker(sync)
 		self.worker.moveToThread(self.thread)
 		self.thread.started.connect(self.worker.run)
+		if finished:
+			self.worker.finished.connect(finished)
 		self.worker.finished.connect(self.thread.quit)
 		self.worker.finished.connect(self.worker.deleteLater)
 		self.thread.finished.connect(self.thread.deleteLater)
 		self.worker.finished.connect(self.threadFinished)
 		self.worker.progress.connect(self.onProgress)
 		self.thread.start()
-		
 
-	# copyLocalFiles 
+	# dumpMemoryCard 
 	# NOTE: if you have two memory card for different sensors they will end up in the same directory
-	def copyLocalFiles(self, drive, num_sensor):
+	def dumpMemoryCard(self, drive, field_name, num_sensor):
 		log.print("*** Dumping usb drive")		
 		
 		dst_dir=LOCAL_DIR
 		dst_dir+='/{}'.format(T1.strftime("%Y%m%d-%H%M%S"))
-		if len(self.field_name.text()):
-			dst_dir+='-{}'.format(self.field_name.text())
+		dst_dir+='-{}'.format(field_name.text())
 		dst_dir+="/sensor{}".format(num_sensor)
-		dst_dir=Utils.NormalizePath(dst_dir) 		
+		dst_dir=Utils.NormalizePath(dst_dir) 	
 		
 		self.setRunning(True)
-		sync=SyncFiles(drive,dst_dir,copier=CopyLocalFiles(), purge=False)
-		self.runSyncInBackground(sync)
+		src_dir=drive
+		
+		copier=CompoundCopyFiles()
+		copier.instances.append(CopyLocalFiles(dst_dir))
+		
+		sync=SyncFiles(src_dir,copier=copier)
+		
+		def onFinished():
+			#Utils.EjectUSBDrive(drive)
+			#self.refreshButtons()	
+			pass
+		
+		self.runSyncInBackground(sync,finished=onFinished)
 
-	# copyFilesToGoogleDrive
-	def copyFilesToGoogleDrive(self):
-		log.print("*** Copying file to google drive")	
-		src_dir=LOCAL_DIR
-		dst_dir=BUCKET_NAME
+	# copyToCloud
+	def copyToCloud(self,drive=None):
+		log.print("*** Copying file to google drive","and copying to drive {}".format(drive) if drive else "")	
+		
 		self.setRunning(True)
-		sync=SyncFiles(src_dir,dst_dir,copier=CopyFilesToGoogleDrive(), purge=False)
+		
+		copier=CompoundCopyFiles()
+		
+		# first copy to google drive
+		copier.instances.append(CopyFilesToGoogleDrive(BUCKET_NAME)) 
+		
+		# eventually copy to external usb drive
+		if drive:
+			copier.instances.append(CopyLocalFiles(BACKUP_DIR.format(drive)))
+		
+		src_dir=LOCAL_DIR
+		sync=SyncFiles(src_dir,copier=copier)
 		self.runSyncInBackground(sync)
